renames package selection to tokenizers
